moved logger into separate file within winston dirA
