+++ config/winston/index.js
@@ -1,34 +1,9 @@
 // enforces good coding habits, such as utilizing all vars
 "use strict";
 
-const winston = require('winston');
-
-const winOptions = 
-    {
-        level: 'info',
-        format: winston.format.json(),
-        defaultMeta: { service: 'user-service' },
-        transports: [
-            new winston.transports.File( 
-                { filename: 'error.log', level: 0}),
-            new winston.transports.File( 
-                { filename: 'combined.log' }),
-            ]
-    }
-
-const logger = winston.createLogger(winOptions);
-
-// ------ debugging section to print verbose logs in console ------ //
-
-logger.add(new winston.transports.Console(
-    {
-        format: winston.format.simple()
-    }
-));
-
-// ------ end debugging                                      ------ // 
+const logger = require('./logger.js');
 
 module.exports = 
-    { 
-        winston
-    };
+{
+    logger
+};
